+++ processor/opentelemetry-processor-baggage/src/opentelemetry/processor/baggage/__init__.py
@@ -13,8 +13,7 @@
 # limitations under the License.
 
 # pylint: disable=import-error
-
 from .processor import ALLOW_ALL_BAGGAGE_KEYS, BaggageSpanProcessor
+from .log_processor import BaggageLogProcessor
 from .version import __version__
-
-__all__ = ["ALLOW_ALL_BAGGAGE_KEYS", "BaggageSpanProcessor", "__version__"]
+__all__ = ["ALLOW_ALL_BAGGAGE_KEYS", "BaggageSpanProcessor", "BaggageLogProcessor", "__version__"]